# scripts/swagger.yaml
openapi: 3.0.0
info:
  title: Doctor Note Summarizer API
  version: 1.1.1
  description: >
    Prototype API for summarizing, redacting, and exporting doctor notes.
    Supports JSON and raw text input. Authentication is handled via Auth0
    (OIDC) plus an Express session cookie.
...
servers:
  - url: http://localhost:3000
security:
  - cookieAuth: []
paths:
...
  /api/audit/logs:
    get:
      summary: Retrieve audit logs (Admin only)
      description: |
        Returns a list of recent audit log entries from the log file.  
        Only users with the **Administrator** role can access this endpoint.
      tags:
        - Administration
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of log entries to return (default 200)
          required: false
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Audit logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 10
                  logs:
                    type: array
                    items:
                      type: object
                      description: Single audit log entry (shape may vary)
        "403":
          description: Access denied (non-admin user)
        "500":
          description: Server error retrieving audit logs
  /api/audit/log:
    post:
      summary: Add a new audit log entry manually (testing)
      tags:
        - Administration
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  example: manual
                message:
                  type: string
                  example: "Manual audit entry from Swagger"
                context:
                  type: string
                  example: "swagger-test"
      responses:
        "200":
          description: Audit entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Audit entry recorded."
        "500":
          description: Failed to create audit entry
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >
        Session-based authentication.  
        This uses the Express session cookie (`connect.sid`).  
        After logging in via `/login` (Auth0), Swagger will send your
        session cookie with subsequent API requests.
...